requirements.txt
absl-py==0.9.0
aist-plusplus-api @ file:///C:/Users/User/AppData/Local/Temp/aistplusplus_api
aniposelib @ git+https://github.com/liruilong940607/aniposelib@6a1c252ec427d7c5c3c95ec50773c4371d80b053
appdirs==1.4.4
asgiref==3.6.0
astunparse==1.6.3
audioread==3.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cycler==0.11.0
Cython==0.29.32
decorator==4.4.2
dill==0.3.6
Django==3.2.16
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
dm-tree==0.1.7
docopt==0.6.2
einops @ file:///C:/Users/User/AppData/Local/Temp/einops
etils==0.9.0
ffmpeg-python==0.2.0
filelock==3.8.0
flatbuffers==22.11.23
fonttools==4.38.0
future==0.18.2
gast==0.4.0
gdown==4.5.4
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
h5py==3.7.0
httplib2==0.21.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
immutabledict==2.2.3
importlib-metadata==5.1.0
importlib-resources==5.10.0
joblib==1.2.0
Js2Py==0.74
kaggle==1.5.12
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
moviepy==1.0.3
mutagen==1.46.0
networkx==2.6.3
numba==0.56.4
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
OpenEXR @ file:///C:/Users/User/pipwin/OpenEXR-1.3.7-cp37-cp37m-win_amd64.whl
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pipwin==0.5.2
pooch==1.6.0
portalocker==2.6.0
proglog==0.1.10
promise==2.3
protobuf==3.19.6
psutil==5.9.4
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==5.4.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.97
seqeval==1.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smplx @ git+https://github.com/liruilong940607/smplx@438732844373ea9ea7799a0d7a467426a674c6fe
soundfile==0.11.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-addons==0.18.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-graphics==2021.12.3
tensorflow-graphics-gpu==1.0.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-metadata==1.11.0
tensorflow-model-optimization==0.7.3
tensorflow-text==2.10.0
termcolor==2.1.1
text-unidecode==1.3
tf-models-official==2.10.1
tf-slim==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.0
torchgeometry==0.1.2
torchvision==0.14.0
tqdm==4.64.1
trimesh==3.17.1
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.11.0
